+++ src/main/java/commands/AddCommand.java
@@ -1,3 +1,14 @@
+package commands;
+import exceptions.DukeException;
+import io.Storage;
+import tasks.DeadlineTask;
+import tasks.EventTask;
+import tasks.Task;
+import tasks.TaskList;
+import tasks.TodoTask;
+import ui.Ui;
+
+
 public class AddCommand extends Command {
 
     public AddCommand(String command) {
