+++ src/main/java/commands/Command.java
@@ -1,3 +1,10 @@
+package commands;
+
+import exceptions.DukeException;
+import io.Storage;
+import tasks.TaskList;
+import ui.Ui;
+
 public abstract class Command {
     private String fullCommand;
     private boolean isExit;
