+++ src/main/java/commands/LoadCommand.java
@@ -1,3 +1,9 @@
+package commands;
+import exceptions.DukeException;
+import io.Storage;
+import tasks.TaskList;
+import ui.Ui;
+
 public class LoadCommand extends Command {
     public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
         storage.load(taskList);
