+++ src/main/java/commands/MarkCommand.java
@@ -1,3 +1,9 @@
+package commands;
+import exceptions.DukeException;
+import io.Storage;
+import tasks.TaskList;
+import ui.Ui;
+
 public class MarkCommand extends Command {
     public MarkCommand(String fullCommand) {
         super(fullCommand);
