+++ src/main/java/exceptions/DukeException.java
@@ -1,3 +1,4 @@
+package exceptions;
 public class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
