+++ text-ui-test/data/tasks.txt
@@ -1,5 +1,6 @@
-E | X | asdlkj | 20/12/2020 1800 | 5/5/2023 1900
+T | X | for
+E |   | asdlkj | 20/12/2020 1800 | 5/5/2023 1900
 D |   |  asad | 20/12/2020 1800
-T |   | salkdjlkj
+T |   | for
 E |   | asdlkj | 20/12/2020 1800 | 5/5/2023 1900
 D |   |  asad | 20/12/2020 1800
